fix (some) build errors for new software versions

+++ imaging/applications/image-accumulate.cpp
@@ -36,7 +36,7 @@
 #include <limits>
 #include <map>
 #include <boost/array.hpp>
-#include <boost/bind.hpp>
+#include <boost/bind/bind.hpp>
 #include <boost/date_time/posix_time/ptime.hpp>
 #include <boost/function.hpp>
 #include <boost/optional.hpp>
@@ -350,7 +350,7 @@ static void output_once_( const boost::posix_time::ptime& t )
 
 static void output_()
 {
-    static const boost::posix_time::time_duration period = boost::posix_time::milliseconds( 1000.0 / *fps );
+    static const boost::posix_time::time_duration period = boost::posix_time::milliseconds( (long) ( 1000.0 / *fps ) );
     static const boost::posix_time::time_duration timeout = boost::posix_time::milliseconds( 10 );
     boost::posix_time::ptime time_to_output = boost::posix_time::microsec_clock::universal_time() + period;    
     while( !is_shutdown && !done && std::cout.good() )
